# frontend/person-blog-app/packages/blog/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Gateway.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/video"
    }
  ],
  "paths": {
...
    "/api/Blog/create": {
      "post": {
        "tags": [
          "Blog"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "PhotoUrl": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "PhotoUrl": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlogModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/BlogPostV2/my": {
      "get": {
        "tags": [
          "BlogPostV2"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "postType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PostType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModelPagedListViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModelPagedListViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModelPagedListViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPostV2/availablePostByBlogId/{blogId}": {
      "get": {
        "tags": [
          "BlogPostV2"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "postType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PostType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPostV2/create": {
      "get": {
        "tags": [
          "BlogPostV2"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePostModelViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePostModelViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePostModelViewModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BlogPostV2"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Visibility": {
                    "$ref": "#/components/schemas/PostVisibility"
                  },
                  "VideoPostData.Thumbnail": {
                    "type": "string",
                    "format": "binary"
                  },
                  "VideoPostData.Description": {
                    "type": "string"
                  },
                  "VideoPostData.Categories": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32"
                    }
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "Visibility": {
                  "style": "form"
                },
                "VideoPostData.Thumbnail": {
                  "style": "form"
                },
                "VideoPostData.Description": {
                  "style": "form"
                },
                "VideoPostData.Categories": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPostV2/remove/{postId}": {
      "post": {
        "tags": [
          "BlogPostV2"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/BlogPostV2/edit/{postId}": {
      "get": {
        "tags": [
          "BlogPostV2"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostEditViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostEditViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostEditViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPostV2/edit": {
      "post": {
        "tags": [
          "BlogPostV2"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "Title": {
                    "minLength": 3,
                    "type": "string"
                  },
                  "Preview": {
                    "type": "string",
                    "format": "binary"
                  },
                  "Categories": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32"
                    }
                  },
                  "Visibility": {
                    "$ref": "#/components/schemas/PostVisibility"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "Preview": {
                  "style": "form"
                },
                "Categories": {
                  "style": "form"
                },
                "Visibility": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/BlogPostV2/createTextPost": {
      "post": {
        "tags": [
          "BlogPostV2"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Title"
                ],
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Text": {
                    "type": "string"
                  },
                  "Visibility": {
                    "$ref": "#/components/schemas/PostVisibility"
                  },
                  "Media": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "InlineMedia": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "InlineMediaIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "Text": {
                  "style": "form"
                },
                "Visibility": {
                  "style": "form"
                },
                "Media": {
                  "style": "form"
                },
                "InlineMedia": {
                  "style": "form"
                },
                "InlineMediaIds": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostInfoModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPostV2/textEdit/{postId}": {
      "get": {
        "tags": [
          "BlogPostV2"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TextPostEditViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextPostEditViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextPostEditViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPostV2/textEdit": {
      "post": {
        "tags": [
          "BlogPostV2"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Title"
                ],
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "Title": {
                    "minLength": 3,
                    "type": "string"
                  },
                  "Text": {
                    "type": "string"
                  },
                  "Visibility": {
                    "$ref": "#/components/schemas/PostVisibility"
                  },
                  "Media": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "InlineMedia": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "InlineMediaIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "RemovedMediaIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "Text": {
                  "style": "form"
                },
                "Visibility": {
                  "style": "form"
                },
                "Media": {
                  "style": "form"
                },
                "InlineMedia": {
                  "style": "form"
                },
                "InlineMediaIds": {
                  "style": "form"
                },
                "RemovedMediaIds": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlogModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "subscribersCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CategoryModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreatePostModelViewModel": {
        "type": "object",
        "properties": {
          "subscriptionLevels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubscriptionLevelModel"
            },
            "nullable": true
          },
          "visibility": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostVisibilitySelectItem"
            },
            "nullable": true
          },
          "categoryList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostCommonModelV2": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "blogId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "postType": {
            "$ref": "#/components/schemas/PostTypeModel"
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PostCommonModelV2PagedListViewModel": {
        "required": [
          "items",
          "pageSize",
          "totalPageCount",
          "totalPostsCount"
        ],
        "type": "object",
        "properties": {
          "totalPageCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPostsCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostCommonModelV2"
            }
          }
        },
        "additionalProperties": false
      },
...
      "PostEditViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "visibility": {
            "$ref": "#/components/schemas/PostVisibility"
          },
          "paymentSubscriptionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "PostTypeModel": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "PostVisibility": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "PostVisibilitySelectItem": {
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/PostVisibility"
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProcessState": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SubscriptionLevelModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "blogId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "nextLevelId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TextInfoDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TextPostEditViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "visibility": {
            "$ref": "#/components/schemas/PostVisibility"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TextPostMediaViewModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TextPostMediaViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "length": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "UserPostInfoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "blogId": {
            "type": "string",
            "format": "uuid"
          },
          "dislikeCount": {
            "type": "integer",
            "format": "int32"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "viewCount": {
            "type": "integer",
            "format": "int64"
          },
          "paymentSubscriptionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "visibility": {
            "$ref": "#/components/schemas/PostVisibility"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "textInfo": {
            "$ref": "#/components/schemas/TextInfoDto"
          },
          "videoInfo": {
            "$ref": "#/components/schemas/VideoInfoDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UserPostInfoModelPagedListViewModel": {
        "required": [
          "items",
          "pageSize",
          "totalPageCount",
          "totalPostsCount"
        ],
        "type": "object",
        "properties": {
          "totalPageCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPostsCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPostInfoModel"
            }
          }
        },
        "additionalProperties": false
      },
...
      "VideoInfoDto": {
        "type": "object",
        "properties": {
          "processState": {
            "$ref": "#/components/schemas/ProcessState"
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "videoMetadata": {
            "$ref": "#/components/schemas/VideoMetadataModel"
          }
        },
        "additionalProperties": false
      },
      "VideoMetadataModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "length": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "objectName": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}